+++ rubato/utils/__init__.py
@@ -10,6 +10,7 @@ from .defaults import Defaults
 from .display import Display
 from .color import Color
 from .font import Font
+from .draw import Draw
 from .time import Time
 from .radio import Radio
 from .input import Input