+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 -   Moved collision test code and impulse resolution to `Engine` class.
 -   Changed default physics fps to 30.
 -   `Polygon.generate_polygon()` can now takes an optional `options` parameter. When set, it returns a `Polygon` instead of a list of vertices.
+-   Move basic draw functions from `Display` to `Draw`.
 
 ### Removed